Data Scientist Path
Files related to Data Scientist Path at Microsoft Learn. Please note these have dependencies and it is good to experiment in the course labs related to this career path.  This is to demonstrates the results and errors if any. Proving the files AS-IS without any modifications. Errors need to be handled properly wherever possible or reach out to the course team.